[Style] #401 - extension 컨벤션 적용
