+++ HappyAnding/HappyAnding/Extensions/Color+Extension.swift
@@ -6,29 +6,6 @@
 //
 
 import SwiftUI
-import UIKit
-
-extension String {
-    
-    func converToColor() -> Color {
-        Color(self)
-    }
-}
-
-extension UIColor {
-    convenience init(light: UIColor, dark: UIColor) {
-        self.init { traitCollection in
-            switch traitCollection.userInterfaceStyle {
-            case .light, .unspecified:
-                return light
-            case .dark:
-                return dark
-            @unknown default:
-                return light
-            }
-        }
-    }
-}
 
 extension Color {
     init(light: Color, dark: Color) {
@@ -37,7 +14,8 @@ extension Color {
 }
 
 extension Color {
-    // Semantic Color
+    
+    // MARK: - Semantic Color
     static let backgroundPlus = Color(light: .gray1, dark: .shortcutsZipWhite)
     static let backgroundTabbar = Color(light: .shortcutsZipWhite, dark: Color("Grey005"))
     static let backgroudList = Color(light: .shortcutsZipWhite, dark: .shortcutsZipBackground)
@@ -58,7 +36,7 @@ extension Color {
 }
 
 extension Color {
-    ///System Colors
+    // MARK: - System Colors
     static let shortcutsZipBackground = Color("Grey020")
     static let shortcutsZipWhite = Color("Grey010")
     static let shortcutsZipPrimary = Color("Primary")
@@ -66,7 +44,7 @@ extension Color {
     static let shortcutsZipDanger = Color("Danger")
     static let shortcutsZipSuccess = Color("Success")
     
-    ///Text Colors
+    // MARK: - Text Colors
     static let gray1 = Color("Grey030")
     static let gray2 = Color("Grey040")
     static let gray3 = Color("Grey050")
@@ -74,8 +52,13 @@ extension Color {
     static let gray5 = Color("Grey070")
     static let gray6 = Color("Grey080")
     
-    //gradient사용하는 곳에서 Color.fetchGradient(color: "Red")와 같이 사용해주세요
-    //컬러명은 숫자를 제외하고 UpperCamelCase로 입력해주세요
+    
+    /**
+     # code
+     
+        Color.fetchGradient(color: 컬러명)
+     
+     */
     static func fetchGradient(color: String) -> LinearGradient {
         let colors = [Color("\(color)01"), Color("\(color)02")]
         return LinearGradient(colors: colors, startPoint: .topLeading, endPoint: .bottomTrailing)
@@ -86,25 +69,3 @@ extension Color {
     }
 }
 
-extension UIColor {
-    static var gray4: UIColor {
-        guard let color = UIColor(named: "Grey060") else { return .label }
-        return color
-    }
-    static var gray5: UIColor {
-        guard let color = UIColor(named: "Grey070") else { return .label }
-        return color
-    }
-    static var shortcutsZipBackground: UIColor {
-        guard let color = UIColor(named: "Grey020") else { return .label }
-        return color
-    }
-    static var shortcutsZipPrimary: UIColor {
-        guard let color = UIColor(named: "Primary") else { return .label }
-        return color
-    }
-    static var shortcutsZipPrimaryOpacity: UIColor {
-        guard let color = UIColor(named: "Primary") else { return .label }
-        return color.withAlphaComponent(0.3)
-    }
-}
