+++ HappyAnding/HappyAnding/Extensions/Image+Extension.swift
@@ -8,6 +8,8 @@
 import SwiftUI
 
 extension Image {
+    
+    /// Color가 gray4, size: 24, frame size가 32인 Cell을 관리하는 함수입니다.
     func setCellIcon() -> some View {
         self
             .foregroundColor(.gray4)
