--- HappyAnding/HappyAnding/Extensions/LineBreaking+Extension.swift
@@ -1,12 +1,0 @@
-//
-//  LineBreaking+Extension.swift
-//  HappyAnding
-//
-//  Created by HanGyeongjun on 2023/01/20.
-//
-
-import Foundation
-
-extension String {
-    var lineBreaking: String { self + "           " }
-}
