+++ HappyAnding/HappyAnding/Extensions/String/String+Formatter.swift
@@ -9,6 +9,10 @@ import SwiftUI
 
 extension String {
     
+    /// SwiftUI의 고질적인 줄바꿈 이슈를 해결하기 위한 변수입니다.
+    var lineBreaking: String { self + "           " }
+    
+    /// 특수문자가 포함되어있는지 확인하는 함수입니다.
     func isNormalString() -> Bool {
         let specialCharacters = "!@#$%^&*()~`_+-=[]{}|;':\",./<>? "
         var result = true
@@ -21,6 +25,7 @@ extension String {
         return result
     }
     
+    /// 적절한 닉네임인지 확인하는 함수입니다.
     func checkCorrectNickname() -> Bool {
         let arr = Array(self)
         let pattern = "^[가-힣ㄱ-ㅎㅏ-ㅣa-zA-Z0-9_]$"
@@ -36,18 +41,6 @@ extension String {
             }
         }
         return true
-    }
-    
-    // substring to string
-    func substring(from: Int, to: Int) -> String {
-        guard from < count, to >= 0, to - from >= 0 else {
-            return ""
-        }
-        
-        let startIndex = index(self.startIndex, offsetBy: from)
-        let endIndex = index(self.startIndex, offsetBy: to + 1)
-        
-        return String(self[startIndex ..< endIndex])
     }
 }
