+++ HappyAnding/HappyAnding/Extensions/UIFont+Extension.swift
@@ -1,0 +1,20 @@
+//
+//  Font+Extension.swift
+//  HappyAnding
+//
+//  Created by 전지민 on 2022/10/21.
+//
+
+import SwiftUI
+
+extension UIFont {
+    static var LargeTitle = UIFont.systemFont(ofSize: 26, weight: .heavy)
+    static var Title1 = UIFont.systemFont(ofSize: 22, weight: .bold)
+    static var Title2 = UIFont.systemFont(ofSize: 20, weight: .bold)
+    static var Headline = UIFont.systemFont(ofSize: 17, weight: .bold)
+    static var Body1 = UIFont.systemFont(ofSize: 17, weight: .regular)
+    static var Body2 = UIFont.systemFont(ofSize: 15, weight: .regular)
+    static var Subtitle = UIFont.systemFont(ofSize: 15, weight: .heavy)
+    static var Footnote = UIFont.systemFont(ofSize: 13, weight: .regular)
+    static var Sb = UIFont.systemFont(ofSize: 15, weight: .semibold)
+}
